managed some routes
